Playlists

+++ app/controllers/home_controller.rb
@@ -34,8 +34,14 @@ class HomeController < ApplicationController
   end
 
   def deleting
-    @user[:library].delete_at(@user[:library].length - 1 - params[:count].to_i)
-    @user.save
+    unless params[:playlist_id]
+      @user[:library].delete_at(@user[:library].length - 1 - params[:count].to_i)
+      @user.save
+    else
+      @playlist = Playlist.find_by_id(params[:playlist_id])
+      @playlist.songs.destroy(Song.find_by_id(params[:song_id].to_i))
+      @playlist.save
+    end
   end
   
   def load
@@ -44,34 +50,56 @@ class HomeController < ApplicationController
 
   def after_load
     song = params[:song]
-    unless song.nil? || params[:singer] == ""
-      File.open(Rails.root.join('public', 'music', song.original_filename.gsub(' ', '_')), 'wb') do |file|
-        file.write(song.read)
-      end
-      new_song = Song.new(name: params[:song].original_filename.delete(".mp3"))
-      singers = params[:singer].gsub(' ft. ', ', ').split(', ')
-      singers.each do |singer|
-        check_singer = Singer.find_by_name(singer)
-        if check_singer.nil?
-          new_singer = Singer.create!(name:singer)
-          new_singer.songs << new_song
-        else
-          check_singer.songs << new_song
+    unless song.original_filename.match('.mp3\z').nil?
+      unless song.nil? || params[:singer] == ""
+        File.open(Rails.root.join('public', 'music', song.original_filename.gsub(' ', '_')), 'wb') do |file|
+          file.write(song.read)
         end
+        new_song = Song.new(name: params[:song].original_filename.delete(".mp3"))
+        singers = params[:singer].gsub(' ft. ', ', ').split(', ')
+        singers.each do |singer|
+          check_singer = Singer.find_by_name(singer)
+          if check_singer.nil?
+            new_singer = Singer.create!(name:singer)
+            new_singer.songs << new_song
+          else
+            check_singer.songs << new_song
+          end
+        end
+        new_song.save
+        redirect_to load_path, notice: 'Загрузка выполнена успешно'
+      else
+        redirect_to load_path, notice: 'Вы не заполнили все условия'
       end
-      new_song.save
-      redirect_to load_path, notice: 'Загрузка выполнена успешно'
     else
-      redirect_to load_path, notice: 'Вы не заполнили все условия'
+        redirect_to load_path, notice: 'Формат должен быть mp3'
     end
   end
 
   def creating_playlist
-    
   end
 
   def creating_playlist_after
-    
+    Playlist.create!(name: params[:name], user: @user)
+    redirect_to library_path
+  end
+
+  def adding_playlist
+  end
+
+  def after_adding_playlist
+    @playlist = Playlist.find_by_id(params[:playlist_id])
+    @playlist.songs << Song.find_by_id(params[:song_id].to_i)
+    @playlist.save
+  end
+
+  def edit_playlist
+    @playlist = Playlist.find_by_id(params[:playlist_id])
+  end
+
+  def deleting_playlist
+    Playlist.find_by_id(params[:playlist_id]).delete
+    redirect_to library_path
   end
   private
